Airflow/requirements.txt
apache-airflow-providers-google==10.25.0
scikit-learn==1.5.2
scipy==1.14.1
numpy==1.26.4
pandas>=2.1.2,<2.2
joblib==1.4.2
mlflow==2.18.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
db-dtypes==1.3.1
pyarrow==18.1.0
requests==2.32.3
